+++ src/main/java/bc1/gream/domain/buy/controller/BuyController.java
@@ -1,9 +1,9 @@
 package bc1.gream.domain.buy.controller;
 
 
-import static bc1.gream.domain.user.coupon.entity.CouponStatus.ALREADY_USED;
-import static bc1.gream.domain.user.coupon.entity.CouponStatus.AVAILABLE;
-import static bc1.gream.domain.user.coupon.entity.CouponStatus.IN_USE;
+import static bc1.gream.domain.coupon.entity.CouponStatus.ALREADY_USED;
+import static bc1.gream.domain.coupon.entity.CouponStatus.AVAILABLE;
+import static bc1.gream.domain.coupon.entity.CouponStatus.IN_USE;
 
 import bc1.gream.domain.buy.dto.request.BuyBidRequestDto;
 import bc1.gream.domain.buy.dto.request.BuyNowRequestDto;
