+++ src/main/java/bc1/gream/domain/buy/service/BuyService.java
@@ -22,10 +22,10 @@ import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.product.service.query.ProductService;
 import bc1.gream.domain.sell.entity.Sell;
 import bc1.gream.domain.sell.repository.SellRepository;
-import bc1.gream.domain.user.coupon.entity.Coupon;
-import bc1.gream.domain.user.coupon.entity.DiscountType;
+import bc1.gream.domain.coupon.entity.Coupon;
+import bc1.gream.domain.coupon.entity.DiscountType;
 import bc1.gream.domain.user.entity.User;
-import bc1.gream.domain.user.service.CouponService;
+import bc1.gream.domain.coupon.service.CouponService;
 import bc1.gream.global.exception.GlobalException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
