+++ src/main/java/bc1/gream/domain/common/facade/ChangingCouponStatusFacade.java
@@ -2,9 +2,9 @@ package bc1.gream.domain.common.facade;
 
 import bc1.gream.domain.buy.entity.Buy;
 import bc1.gream.domain.buy.service.BuyService;
-import bc1.gream.domain.user.coupon.entity.CouponStatus;
+import bc1.gream.domain.coupon.entity.CouponStatus;
 import bc1.gream.domain.user.entity.User;
-import bc1.gream.domain.user.service.CouponService;
+import bc1.gream.domain.coupon.service.CouponService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
