+++ src/main/java/bc1/gream/domain/coupon/entity/CouponStatus.java
@@ -1,4 +1,4 @@
-package bc1.gream.domain.user.coupon.entity;
+package bc1.gream.domain.coupon.entity;
 
 public enum CouponStatus {
     AVAILABLE,
