+++ src/main/java/bc1/gream/domain/order/service/command/OrderCommandService.java
@@ -3,8 +3,8 @@ package bc1.gream.domain.order.service.command;
 import bc1.gream.domain.buy.entity.Buy;
 import bc1.gream.domain.order.entity.Order;
 import bc1.gream.domain.order.repository.OrderRepository;
-import bc1.gream.domain.user.coupon.entity.Coupon;
-import bc1.gream.domain.user.coupon.helper.CouponCalculator;
+import bc1.gream.domain.coupon.entity.Coupon;
+import bc1.gream.domain.coupon.helper.CouponCalculator;
 import bc1.gream.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
