+++ src/main/java/bc1/gream/domain/sell/repository/helper/SellQueryOrderFactory.java
@@ -3,7 +3,7 @@ package bc1.gream.domain.sell.repository.helper;
 
 import static bc1.gream.domain.sell.entity.QSell.sell;
 
-import bc1.gream.domain.product.repository.helper.QueryDslUtil;
+import bc1.gream.global.querydsl.QueryDslUtil;
 import com.querydsl.core.types.Order;
 import com.querydsl.core.types.OrderSpecifier;
 import java.util.ArrayList;
