+++ src/main/java/bc1/gream/domain/sell/service/SellNowService.java
@@ -8,9 +8,9 @@ import bc1.gream.domain.order.mapper.OrderMapper;
 import bc1.gream.domain.order.service.command.OrderCommandService;
 import bc1.gream.domain.sell.dto.request.SellNowRequestDto;
 import bc1.gream.domain.sell.dto.response.SellNowResponseDto;
-import bc1.gream.domain.user.coupon.entity.Coupon;
+import bc1.gream.domain.coupon.entity.Coupon;
 import bc1.gream.domain.user.entity.User;
-import bc1.gream.domain.user.service.CouponService;
+import bc1.gream.domain.coupon.service.CouponService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
