+++ src/test/java/bc1/gream/domain/order/repository/BuyRepositoryTest.java
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import bc1.gream.domain.buy.entity.Buy;
 import bc1.gream.domain.buy.repository.BuyRepository;
 import bc1.gream.domain.product.repository.ProductRepository;
-import bc1.gream.domain.user.coupon.repository.CouponRepository;
+import bc1.gream.domain.coupon.repository.CouponRepository;
 import bc1.gream.domain.user.repository.UserRepository;
 import bc1.gream.global.common.ResultCase;
 import bc1.gream.global.config.QueryDslConfig;
